# kafka.writeBuffer