--- a/fetch.go
+++ b/fetch.go
@@ -1,7 +1,5 @@
 package kafka
 
-import "bufio"
-
 type fetchRequestV2 struct {
 	ReplicaID   int32
 	MaxWaitTime int32
@@ -13,11 +11,11 @@ func (r fetchRequestV2) size() int32 {
 	return 4 + 4 + 4 + sizeofArray(len(r.Topics), func(i int) int32 { return r.Topics[i].size() })
 }
 
-func (r fetchRequestV2) writeTo(w *bufio.Writer) {
-	writeInt32(w, r.ReplicaID)
-	writeInt32(w, r.MaxWaitTime)
-	writeInt32(w, r.MinBytes)
-	writeArray(w, len(r.Topics), func(i int) { r.Topics[i].writeTo(w) })
+func (r fetchRequestV2) writeTo(wb *writeBuffer) {
+	wb.writeInt32(r.ReplicaID)
+	wb.writeInt32(r.MaxWaitTime)
+	wb.writeInt32(r.MinBytes)
+	wb.writeArray(len(r.Topics), func(i int) { r.Topics[i].writeTo(wb) })
 }
 
 type fetchRequestTopicV2 struct {
@@ -30,9 +28,9 @@ func (t fetchRequestTopicV2) size() int32 {
 		sizeofArray(len(t.Partitions), func(i int) int32 { return t.Partitions[i].size() })
 }
 
-func (t fetchRequestTopicV2) writeTo(w *bufio.Writer) {
-	writeString(w, t.TopicName)
-	writeArray(w, len(t.Partitions), func(i int) { t.Partitions[i].writeTo(w) })
+func (t fetchRequestTopicV2) writeTo(wb *writeBuffer) {
+	wb.writeString(t.TopicName)
+	wb.writeArray(len(t.Partitions), func(i int) { t.Partitions[i].writeTo(wb) })
 }
 
 type fetchRequestPartitionV2 struct {
@@ -45,10 +43,10 @@ func (p fetchRequestPartitionV2) size() int32 {
 	return 4 + 8 + 4
 }
 
-func (p fetchRequestPartitionV2) writeTo(w *bufio.Writer) {
-	writeInt32(w, p.Partition)
-	writeInt64(w, p.FetchOffset)
-	writeInt32(w, p.MaxBytes)
+func (p fetchRequestPartitionV2) writeTo(wb *writeBuffer) {
+	wb.writeInt32(p.Partition)
+	wb.writeInt64(p.FetchOffset)
+	wb.writeInt32(p.MaxBytes)
 }
 
 type fetchResponseV2 struct {
@@ -60,9 +58,9 @@ func (r fetchResponseV2) size() int32 {
 	return 4 + sizeofArray(len(r.Topics), func(i int) int32 { return r.Topics[i].size() })
 }
 
-func (r fetchResponseV2) writeTo(w *bufio.Writer) {
-	writeInt32(w, r.ThrottleTime)
-	writeArray(w, len(r.Topics), func(i int) { r.Topics[i].writeTo(w) })
+func (r fetchResponseV2) writeTo(wb *writeBuffer) {
+	wb.writeInt32(r.ThrottleTime)
+	wb.writeArray(len(r.Topics), func(i int) { r.Topics[i].writeTo(wb) })
 }
 
 type fetchResponseTopicV2 struct {
@@ -75,9 +73,9 @@ func (t fetchResponseTopicV2) size() int32 {
 		sizeofArray(len(t.Partitions), func(i int) int32 { return t.Partitions[i].size() })
 }
 
-func (t fetchResponseTopicV2) writeTo(w *bufio.Writer) {
-	writeString(w, t.TopicName)
-	writeArray(w, len(t.Partitions), func(i int) { t.Partitions[i].writeTo(w) })
+func (t fetchResponseTopicV2) writeTo(wb *writeBuffer) {
+	wb.writeString(t.TopicName)
+	wb.writeArray(len(t.Partitions), func(i int) { t.Partitions[i].writeTo(wb) })
 }
 
 type fetchResponsePartitionV2 struct {
@@ -92,10 +90,10 @@ func (p fetchResponsePartitionV2) size() int32 {
 	return 4 + 2 + 8 + 4 + p.MessageSet.size()
 }
 
-func (p fetchResponsePartitionV2) writeTo(w *bufio.Writer) {
-	writeInt32(w, p.Partition)
-	writeInt16(w, p.ErrorCode)
-	writeInt64(w, p.HighwaterMarkOffset)
-	writeInt32(w, p.MessageSetSize)
-	p.MessageSet.writeTo(w)
+func (p fetchResponsePartitionV2) writeTo(wb *writeBuffer) {
+	wb.writeInt32(p.Partition)
+	wb.writeInt16(p.ErrorCode)
+	wb.writeInt64(p.HighwaterMarkOffset)
+	wb.writeInt32(p.MessageSetSize)
+	p.MessageSet.writeTo(wb)
 }