--- a/joingroup.go
+++ b/joingroup.go
@@ -24,17 +24,15 @@ func (t groupMetadata) size() int32 {
 		sizeofBytes(t.UserData)
 }
 
-func (t groupMetadata) writeTo(w *bufio.Writer) {
-	writeInt16(w, t.Version)
-	writeStringArray(w, t.Topics)
-	writeBytes(w, t.UserData)
+func (t groupMetadata) writeTo(wb *writeBuffer) {
+	wb.writeInt16(t.Version)
+	wb.writeStringArray(t.Topics)
+	wb.writeBytes(t.UserData)
 }
 
 func (t groupMetadata) bytes() []byte {
 	buf := bytes.NewBuffer(nil)
-	w := bufio.NewWriter(buf)
-	t.writeTo(w)
-	w.Flush()
+	t.writeTo(&writeBuffer{w: buf})
 	return buf.Bytes()
 }
 
@@ -61,9 +59,9 @@ func (t joinGroupRequestGroupProtocolV1) size() int32 {
 		sizeofBytes(t.ProtocolMetadata)
 }
 
-func (t joinGroupRequestGroupProtocolV1) writeTo(w *bufio.Writer) {
-	writeString(w, t.ProtocolName)
-	writeBytes(w, t.ProtocolMetadata)
+func (t joinGroupRequestGroupProtocolV1) writeTo(wb *writeBuffer) {
+	wb.writeString(t.ProtocolName)
+	wb.writeBytes(t.ProtocolMetadata)
 }
 
 type joinGroupRequestV1 struct {
@@ -98,13 +96,13 @@ func (t joinGroupRequestV1) size() int32 {
 		sizeofArray(len(t.GroupProtocols), func(i int) int32 { return t.GroupProtocols[i].size() })
 }
 
-func (t joinGroupRequestV1) writeTo(w *bufio.Writer) {
-	writeString(w, t.GroupID)
-	writeInt32(w, t.SessionTimeout)
-	writeInt32(w, t.RebalanceTimeout)
-	writeString(w, t.MemberID)
-	writeString(w, t.ProtocolType)
-	writeArray(w, len(t.GroupProtocols), func(i int) { t.GroupProtocols[i].writeTo(w) })
+func (t joinGroupRequestV1) writeTo(wb *writeBuffer) {
+	wb.writeString(t.GroupID)
+	wb.writeInt32(t.SessionTimeout)
+	wb.writeInt32(t.RebalanceTimeout)
+	wb.writeString(t.MemberID)
+	wb.writeString(t.ProtocolType)
+	wb.writeArray(len(t.GroupProtocols), func(i int) { t.GroupProtocols[i].writeTo(wb) })
 }
 
 type joinGroupResponseMemberV1 struct {
@@ -118,9 +116,9 @@ func (t joinGroupResponseMemberV1) size() int32 {
 		sizeofBytes(t.MemberMetadata)
 }
 
-func (t joinGroupResponseMemberV1) writeTo(w *bufio.Writer) {
-	writeString(w, t.MemberID)
-	writeBytes(w, t.MemberMetadata)
+func (t joinGroupResponseMemberV1) writeTo(wb *writeBuffer) {
+	wb.writeString(t.MemberID)
+	wb.writeBytes(t.MemberMetadata)
 }
 
 func (t *joinGroupResponseMemberV1) readFrom(r *bufio.Reader, size int) (remain int, err error) {
@@ -160,13 +158,13 @@ func (t joinGroupResponseV1) size() int32 {
 		sizeofArray(len(t.MemberID), func(i int) int32 { return t.Members[i].size() })
 }
 
-func (t joinGroupResponseV1) writeTo(w *bufio.Writer) {
-	writeInt16(w, t.ErrorCode)
-	writeInt32(w, t.GenerationID)
-	writeString(w, t.GroupProtocol)
-	writeString(w, t.LeaderID)
-	writeString(w, t.MemberID)
-	writeArray(w, len(t.Members), func(i int) { t.Members[i].writeTo(w) })
+func (t joinGroupResponseV1) writeTo(wb *writeBuffer) {
+	wb.writeInt16(t.ErrorCode)
+	wb.writeInt32(t.GenerationID)
+	wb.writeString(t.GroupProtocol)
+	wb.writeString(t.LeaderID)
+	wb.writeString(t.MemberID)
+	wb.writeArray(len(t.Members), func(i int) { t.Members[i].writeTo(wb) })
 }
 
 func (t *joinGroupResponseV1) readFrom(r *bufio.Reader, size int) (remain int, err error) {